requirements.txt
amqp==5.0.6
apns2==0.7.1
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.17.58
botocore==1.20.112
CacheControl==0.12.10
cachetools==4.2.4
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.4
django-admin-rangefilter==0.8.3
django-admin-sortable2==1.0
django-celery-beat==2.2.1
django-filter==2.4.0
django-google-maps==0.12.4
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-push-notifications==2.0.0
django-solo==1.2.0
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
fcm-django==1.0.1
firebase-admin==5.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.20.2
hyper==0.7.0
hyperframe==3.2.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
phonenumbers==8.12.36
Pillow==8.2.0
prompt-toolkit==3.0.20
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywebpush==1.14.0
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
django-mapbox-location-field==1.7.3
django-map-widgets